02-rag-evaluation/pyproject.toml
[project]
name = "02-rag-evaluation"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "braintrust>=0.5.6",
    "cohere>=5.20.5",
    "faiss-cpu>=1.13.2",
    "instructor>=1.14.5",
    "judges>=0.1.1",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=1.1.0",
    "litellm>=1.81.11",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "openai>=2.20.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.54.0",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.1",
]

# Adds project root to sys.path so `from src.models import ...` works in tests
[tool.pytest.ini_options]
pythonpath = ["."]
